Move functions to internals.expr.funcs, add PHI-function
